+++ unicorn_binance_websocket_api/unicorn_binance_websocket_api_connection.py
@@ -200,19 +200,21 @@ class BinanceWebSocketApiConnection(object):
         try:
             await self._conn.__aexit__(*args, **kwargs)
         except AttributeError as error_msg:
-            logging.debug("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
+            logging.error("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
                           "AttributeError - " + str(error_msg))
-            try:
-                if self.handler_binance_websocket_api_manager.websocket_list[self.stream_id]:
-                    self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].close()
-                logging.debug("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
-                              "AttributeError - close() - done!")
-            except KeyError as error_msg:
-                logging.debug("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
-                              "KeyError - " + str(error_msg))
-            except RuntimeWarning as error_msg:
-                logging.debug("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
-                              "RuntimeWarning - " + str(error_msg))
+            # TODO: obsolete?
+            #try:
+            #    self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].close()
+            #    logging.debug("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
+            #                  "AttributeError - close() - done!")
+            #except KeyError as error_msg:
+            #    logging.debug("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
+            #                  "KeyError - " + str(error_msg))
+            #except RuntimeWarning as error_msg:
+            #    logging.debug("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
+            #                  "RuntimeWarning - " + str(error_msg))
+            #except AttributeError:
+            #    sys.exit(1)
         except websockets.exceptions.ConnectionClosed as error_msg:
             logging.critical("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
                              "ConnectionClosed - " + str(error_msg))