+++ example_stream_everything.py
@@ -100,4 +100,5 @@ for channel in channels:
 
 while True:
     binance_websocket_api_manager.print_summary()
-    time.sleep(1)
+    binance_websocket_api_manager.print_summary_to_png("/home/oliver/")
+    time.sleep(2)