+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Close WS only if open in connection class row 190 [issue #72](https://github.com/oliver-zehentleitner/unicorn-binance-websocket-api/issues/72)
 ### Removed
 - some code in connection row 206 which is not needed anymore and is causing a coroutine error
+- `is_websocket_uri_length_valid()`
 
 ## 1.14.0
 ### Added