+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to 
 [Semantic Versioning](http://semver.org/).
 
-## 1.14.0.dev (development stage/unreleased/unstable)
+## 1.15.0.dev (development stage/unreleased/unstable)
+
+## 1.15.0
 ### Added 
 - psutil (new requirement)
 - exception handling of `websockets.exceptions.InvalidMessage` [issue #72](https://github.com/oliver-zehentleitner/unicorn-binance-websocket-api/issues/72)